server/pyproject.toml
[project]
name = "agent-port"
version = "0.1.0"
description = "Universal tool gateway for AI agents"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    "httpx>=0.27.0",
    "mcp>=1.9.0",
    "pydantic-settings>=2.6.0",
    "pyjwt>=2.12.1",
    "python-multipart>=0.0.24",
    "bcrypt>=5.0.0",
    "python-dotenv>=1.2.2",
    "aiofiles>=24.1.0",
    "resend>=2.28.1",
    "cryptography>=43.0.0",
    "posthog>=7.12.0",
    "pyotp>=2.9.0",
    "qrcode[pil]>=8.2",
    "stripe>=10.0.0",
]

[project.optional-dependencies]
# Install exactly one of these depending on environment:
#   dev/local:  uv sync --extra postgres-binary
#   production: uv sync --extra postgres  (compiled against system libpq)
postgres = ["psycopg2>=2.9.0"]
postgres-binary = ["psycopg2-binary>=2.9.0"]
aws = ["boto3>=1.34.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_port"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
